started refactoring views in chat
